requirements.txt
Django==1.11.1
django-nose==1.4.5
django-rest-framework==0.1.0
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.0
inflection==0.3.1
nose==1.3.7
pytz==2017.2
simplejson==3.10.0
